Add documentation to the data source section.
