+++ ResearchSuite/ResearchSuite/RSDPickerDataSource.swift
@@ -34,70 +34,93 @@
 import Foundation
 import UIKit
 
+/// `RSDPickerDataSource` includes information that can be used to build a picker UI element.
 public protocol RSDPickerDataSource {
 }
 
+/// `RSDDatePickerMode` describes the type of UI picker to display for dates and times.
+/// - seealso: `RSDDatePickerDataSource`
 public enum RSDDatePickerMode : String {
-    case time // Displays hour, minute, and optionally AM/PM designation depending on the locale setting (e.g. 6 | 53 | PM)
     
-    case date // Displays month, day, and year depending on the locale setting (e.g. November | 15 | 2007)
+    /// Displays hour, minute, and optionally AM/PM designation depending on the locale setting (e.g. 6 | 53 | PM)
+    case time
     
-    case dateAndTime // Displays date, hour, minute, and optionally AM/PM designation depending on the locale setting (e.g. Wed Nov 15 | 6 | 53 | PM)
+    /// Displays month, day, and year depending on the locale setting (e.g. November | 15 | 2007)
+    case date
+    
+    /// Displays date, hour, minute, and optionally AM/PM designation depending on the locale setting
+    /// (e.g. Wed Nov 15 | 6 | 53 | PM)
+    case dateAndTime
 }
 
+/// A data source for selecting a date.
 public protocol RSDDatePickerDataSource : RSDPickerDataSource {
     
+    /// The type of UI picker to display for dates and times.
     var datePickerMode: RSDDatePickerMode { get }
     
-    // specify min/max date range. default is nil. When min > max, the values are ignored.
+    /// Specify the minimum date range. Default = `nil`. When `minimumDate` > `maximumDate`, the values are ignored.
     var minimumDate: Date? { get }
+    
+    /// Specify the maximum date range. Default = `nil`. When `minimumDate` > `maximumDate`, the values are ignored.
     var maximumDate: Date? { get }
     
-    // display minutes wheel with interval. interval must be evenly divided into 60. default is 1. min is 1, max is 30
+    /// display minutes wheel with interval. interval must be evenly divided into 60. default is 1. min is 1, max is 30
     var minuteInterval: Int? { get }
     
-    // The date formatter for displaying the date in a text field
+    /// The date formatter for displaying the date in a text field or label
     var dateFormatter: DateFormatter { get }
 }
 
+/// A picker data source for selecting date components.
 public protocol RSDDateComponentPickerDataSource : RSDPickerDataSource {
     
+    /// The calendar to use for the date components
     var calendar: Calendar { get }
+    
+    /// The components to include in the picker
     var calendarComponents: Set<Calendar.Component> { get }
     
-    // Specify min/max year range if the year is included, otherwise these values are ignored.
+    /// The minimum year if the year is included, otherwise this value is ignored.
     var minimumYear: Int? { get }
+    
+    /// The maximum year if the year is included, otherwise this value is ignored.
     var maximumYear: Int? { get }
     
+    /// The date components formatter for displaying the date components in a text field or label
     var dateComponentsFormatter: DateComponentsFormatter { get }
 }
 
+/// A picker data source for selecting choices.
 public protocol RSDChoicePickerDataSource : RSDPickerDataSource {
     
-    // returns the number of 'columns' to display.
+    /// Returns the number of 'columns' to display.
     var numberOfComponents: Int { get }
     
-    // returns the # of rows in each component
+    /// Returns the # of rows in each component.
+    /// - parameter component: The component (or column) of the picker.
+    /// - returns: The number of rows in the given component.
     func numberOfRows(in component: Int) -> Int
     
-    // returns the choice for this row/component. If this is returns `nil` then this is the "optional" choice.
+    /// Returns the choice for this row/component. If this is returns `nil` then this is the "skip" choice.
+    /// - parameters:
+    ///     - row: The row for the selected component.
+    ///     - component: The component (or column) of the picker.
     func choice(forRow row: Int, forComponent component: Int) -> RSDChoice?
-    
-    // returns the selected answer for a given row and componet
-    func selectedAnswer(for selection: [Int : Int]) -> Any
 }
 
-public protocol RSDDecimalPickerDataSource : RSDPickerDataSource {
+/// A picker data source for picking a number.
+public protocol RSDNumberPickerDataSource : RSDPickerDataSource {
     
-    // Returns the minimum number allowed.
+    /// Returns the minimum number allowed.
     var minimum: Decimal { get }
     
-    // Returns the maximum number allowed.
+    /// Returns the maximum number allowed.
     var maximum: Decimal { get }
     
-    // Returns the step interval to use. If `nil`, then the step interval will default to advance by 1.
+    /// Returns the step interval to use. If `nil`, then the step interval will default to advance by 1.
     var stepInterval: Decimal? { get }
     
-    // Returns the number formatter to use to format the displayed value and to parse the result.
+    /// Returns the number formatter to use to format the displayed value and to parse the result.
     var numberFormatter: NumberFormatter { get }
 }
